Changed s3up filename to s3up.php
This is so that IDEs like the one I'm using, PDT, can do syntax highlighting on the file. If you want to be able to type "s3up" as the command, it is a simple matter to just created a symbolic link called "s3up" in your "bin" directory pointing to the php file.

Signed-off-by: Tyler Hall <tylerhall@gmail.com>