Reformatted the script
Changed lines like:
1 if($value)
2 {
3	do();
4 }
5 else
6 {
7	dothis();
8 }

to this:
1 if($value){
2	do();
3 }else{
4	dothis();
5 }

While the first stype used to be necessary, ever since editors added code folding, indentation guides and brace matching
this method has become redundant and now only serves to make the programmer scroll more. For one statement this isn't
such a big deal, but for an entire script or library it becomes annoying.


Changed lines like this:
if($value)
	do();

or this:
if($value) do();

To this:
if($value){
	do();
}

While this uses up more space in the past I have found that as code grows and becomes more complex statements like this
can lead to developers misunderstanding code flow and thus introducing bugs. Also with statements such as
if($file == '.' || $file == '..') continue;
because the continue is not the first thing on the line it is very easy to miss the change in flow of execution when trying to
debug the code.

Amendment: I also changed spaces to tabs as using a single tab for an indent allows the programmer to set indent size to
whatever their preference is as the optimal setting tends to be quite unique to each programmer so enforcing a size for indents
would be forcing your readability preferences onto them when this can simply be an editor preference.

Signed-off-by: Tyler Hall <tylerhall@gmail.com>