Bug fix, correcting user errors and making loops efficent
Bug fix: As the documentation doesn't specify how to use the arguments --key and --secret
(or even that they exist) you have to read the code to figure out how it works. However
in doing so you are lead to believe...
"// Does not have an =, so choose the next arg as its value"
...that you can use either the argument format:
--key=<thekeyvalue>
OR
--key <thekeyvalue>
However due to two bugs in the code trying to do the second version will lead to some very
confusing errors and the accidental public release of your of your secure secret. In this release
I have fixed both bugs.

Correcting user errors: There are a lot of die statements in the original code which is no longer
the best way to kill the script on an error condition in PHP, so I have changed these to instead
call trigger_error with E_USER_ERROR as the status.

Making loops efficent: Two of the loops were doing the same recalculation for every itteration,
however the actual result was static so should only be calculated once before the loops start.

Signed-off-by: Tyler Hall <tylerhall@gmail.com>